refactor(JSONLocalStorage): extend the json local storage helper util so it can has a sessionStorage helper too
