+++ src/utils/index.ts
@@ -3,20 +3,6 @@ import { isSameSecond, parseISO } from 'date-fns';
 // types
 import { UserProfile } from '../@types';
 
-//  Helper function to get auth tokens from query params.
-export const JSONLocalStorage = {
-  get(key: string) {
-    const value = localStorage.getItem(key);
-    return value ? JSON.parse(value) : null;
-  },
-  set(key: string, value: any) {
-    localStorage.setItem(key, JSON.stringify(value));
-  },
-  clear() {
-    localStorage.clear();
-  },
-};
-
 // Helper function to scroll to an element.
 export function scrollToElement(element: HTMLElement) {
   if (typeof element?.scrollIntoView === 'function') {
